Added check and help commands for all functions

+++ personal-env.functions.bash
@@ -4,6 +4,7 @@ COLOR_RED='\033[0;31m'
 COLOR_GREEN='\033[0;32m'
 COLOR_YELLOW='\033[1;33m'
 COLOR_LBLUE='\033[1;34m'
+COLOR_WHITE='\033[0;37m'
 COLOR_NONE='\033[0m'
 
 function __checkIfRoot() {
@@ -42,6 +43,10 @@ function __printInBlue() {
 	__printInColor "$1" $COLOR_LBLUE
 }
 
+function __printInWhite() {
+	__printInColor "$1" $COLOR_WHITE
+}
+
 # stashes the uncommited code, pulls from origin then unstashes
 # Warning if there are conflicts it will not show in interactive UI
 function branchUpdate() {
@@ -49,15 +54,15 @@ function branchUpdate() {
 		unCommittedChanges="Changes not staged for commit:"
 
 		if git status | grep "$unCommittedChanges" ; then
-			echo -e "\nSaving changes to stash\n" && \
+			__printInYellow "\nSaving changes to stash\n" && \
 			git stash save && \
-			echo -e "\nPulling from tracking repo..\n" && \
+			__printInYellow "\nPulling from tracking repo..\n" && \
 			$1 && \
-			echo -e "\nUnstashing changes from stash\n" && \
+			__printInYellow "\nUnstashing changes from stash\n" && \
 			git stash pop
 		else
-			echo -e "\nNo local changes found.." && \
-			echo -e "\nPulling from tracking repo..\n" && \
+			__printInGreen "\nNo local changes found.." && \
+			__printInGreen "\nPulling from tracking repo..\n" && \
 			$1
 		fi
 	}
@@ -80,26 +85,56 @@ function logDiffBetweenBranches() {
 }
 
 function handBrakeConvert() {
-	HandBrakeCLI -i "$1" -o "${2:-$(echo "$1" | sed -e 's/\(^.*[sS][0-9][0-9][eE][0-9][0-9]\)\(.*\)/\1/').mp4}" -e x264 -q 21 -B 160 && \
-	rm "$1"
+	if [ "$#" -eq 1 ] || [ "$#" -eq 2 ]
+	then
+		HandBrakeCLI -i "$1" -o "${2:-$(echo "$1" | sed -e 's/\(^.*[sS][0-9][0-9][eE][0-9][0-9]\)\(.*\)/\1/').mp4}" -e x264 -q 21 -B 160 && \
+		rm "$1"
+	else
+		__printInRed "Unable to execute the command."
+		__printInWhite "Usage: handBrakeConvert source [destination]"
+	fi
 }
 
 function convertAndRmAllInFolder() {
-	for file in $(find "${2:-./}" -type f -name "*.${1:-mkv}"); do
-		handBrakeConvert "$file"
-	done
+	if [ "$#" -eq 1 ] || [ "$#" -eq 2 ]
+	then
+		for file in $(find "${1:-./}" -type f -name "*.${2:-*}"); do
+			handBrakeConvert "$file"
+		done
+	else
+		__printInRed "Unable to execute the command."
+		__printInWhite "Usage: convertAndRmAllInFolder folder_path [extension_to_convert]"
+	fi
 }
 
 function pullSSLCert() {
-	openssl s_client -showcerts -connect "${1}":443 < /dev/null 2> /dev/null | openssl x509 -noout -text
+	if [ "$#" -eq 1 ]
+	then
+		openssl s_client -showcerts -connect "${1}":443 < /dev/null 2> /dev/null | openssl x509 -noout -text
+	else
+		__printInRed "Unable to execute the command."
+		__printInWhite "Usage: pullSSLCert webaddress"	
+	fi
 }
 
 # Get keys from mit liscence keys
 function getGPGKey() {
-	gpg --recv-keys --keyserver hkp://pgp.mit.edu $1
+	if [ "$#" -eq 1 ]
+	then
+		gpg --recv-keys --keyserver hkp://pgp.mit.edu $1
+	else
+		__printInRed "Unable to execute the command."
+		__printInWhite "Usage: getGPGKey keyID"
+	fi
 }
 
 function getCertKeyPemFromPfx() {
-	openssl pkcs12 -in "$1.pfx" -out "$1-cert.pem" -clcerts -nokeys
-    openssl pkcs12 -in "$1.pfx" -out "$1-key.pem" -nocerts -nodes
+	if [ "$#" -eq 1 ]
+	then
+		openssl pkcs12 -in "$1.pfx" -out "$1-cert.pem" -clcerts -nokeys
+    	openssl pkcs12 -in "$1.pfx" -out "$1-key.pem" -nocerts -nodes
+	else
+		__printInRed "Unable to execute the command."
+		__printInWhite "Usage: getCertKeyPemFromPfx name_of_the_pfx"
+	fi
 }
